Added miscellaneous new multivariate theorems:
        CLOSURE_FINITE_INTERS_CONVEX_OPEN
        CLOSURE_INTER_CONVEX
        CLOSURE_INTER_CONVEX_OPEN
        COLLINEAR_EXTREME_POINTS
        COMPACT_CONVEX_COLLINEAR_SEGMENT
        CONVEX_LINEAR_PREIMAGE
        REALLIM_ABS
        REALLIM_MAX
        REALLIM_MIN
        REAL_CONTINUOUS_ABS
        REAL_CONTINUOUS_MAX
        REAL_CONTINUOUS_MIN
        RELATIVE_INTERIOR_OPEN
        SEGMENT_CLOSED_OPEN

The proof of COMPACT_CONVEX_COLLINEAR_SEGMENT uses Krein-Milman,
which is overkill for such a simple result (and the proof wasn't
even all that clean). It might be worth replacing this with a
more direct proof so that it can be used without pulling in the
polytope.ml file, just as I did with EXTREME_POINT_OF_SEGMENT,
whose proof is trivial from Krein-Milman.



git-svn-id: http://hol-light.googlecode.com/svn@37 1eddc5a6-173b-11df-bff1-5f4ea15f2b7c