Made a few changes to miz3: one small fix to the code, removal of the automatic run of "Samples" at the end and the removal of the old "Samples/ALL" which has become the new "test.ml". Also changed "holtest" to run the "test.ml" file twice in case cacheing makes a difference.
Made a fix to the Makefile in response to a bug report from
Bill Richter trying to build with camlp5 6.05. Fortunately
the pa_j_3.1x_6.02.2.ml still works, so it was only necessary
to change the Makefile to use it.

Added a simple but handy set theory lemma INFINITE_SUPERSET.

Fixed some date errors in the CHANGES file itself, pointed
out by Norbert Voelker.



git-svn-id: http://hol-light.googlecode.com/svn@136 1eddc5a6-173b-11df-bff1-5f4ea15f2b7c